requirements.txt
openai>=1.0.0
click>=8.0.0
pyyaml>=6.0
rich>=13.0
black>=24.0
flake8>=7.0
pytest>=8.0
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
sse-starlette>=2.1.0
markdown>=3.6
python-multipart>=0.0.9
python-dotenv>=1.0.0
